+++ webpack.config.cli.prod.js
@@ -9,20 +9,43 @@ const buildBaseConfig = require('./webpack.config.base');
 const cli = require('./webpack.config.cli');
 const esbuildLoaderRule = require('./webpack/esbuild-loader-rule');
 const { getModulesPaths, getStripesModulesPaths, getTranspiledModules } = require('./webpack/module-paths');
+const { addHostMFConfig } = require('./module-federation-config');
 
-const buildConfig = (stripesConfig) => {
+const buildConfig = (stripesConfig, options = {}) => {
   const modulePaths = getModulesPaths(stripesConfig.modules);
   const stripesModulePaths = getStripesModulesPaths();
   const allModulePaths = [...stripesModulePaths, ...modulePaths];
-  const base = buildBaseConfig(allModulePaths);
-  const prodConfig = Object.assign({}, base, cli, {
+  const base = buildBaseConfig(allModulePaths, stripesConfig);
+  const prodOverrides = {
     mode: 'production',
     devtool: 'source-map',
     infrastructureLogging: {
       appendOnly: true,
       level: 'warn',
     },
-  });
+  };
+  let prodConfig = { ...base, ...cli, ...prodOverrides };
+
+  const splitChunks = {
+    // Do not process stripes chunk
+    chunks: (chunk) => {
+      return chunk.name !== 'stripes';
+    },
+    cacheGroups: {
+      // this cache group will be omitted by minimizer
+      stripes: {
+        // only include already transpiled modules
+        test: (module) => transpiledModulesRegex.test(module.resource),
+        name: 'stripes',
+        chunks: 'all'
+      },
+    },
+  };
+
+  if (options.lazy) {
+    splitChunks.chunks = 'all';
+    splitChunks.cacheGroups = undefined;
+  }
 
   const transpiledModules = getTranspiledModules(allModulePaths);
   const transpiledModulesRegex = new RegExp(transpiledModules.join('|'));
@@ -40,21 +63,16 @@ const buildConfig = (stripesConfig) => {
         css: true,
       }),
     ],
-    splitChunks: {
-      // Do not process stripes chunk
-      chunks: (chunk) => {
-        return chunk.name !== 'stripes';
-      },
-      cacheGroups: {
-        // this cache group will be omitted by minimizer
-        stripes: {
-          // only include already transpiled modules
-          test: (module) => transpiledModulesRegex.test(module.resource),
-          name: 'stripes',
-          chunks: 'all'
-        },
-      },
-    },
+    splitChunks,
+  }
+
+  // build platform with Module Federation if --federate flag is passed
+  if (options.federate) {
+    prodConfig.optimization = {
+      concatenateModules: false,
+      ...prodConfig.optimization,
+    };
+    prodConfig = addHostMFConfig(prodConfig);
   }
 
   prodConfig.module.rules.push(esbuildLoaderRule(allModulePaths));